rename config from L1 nomenclature to BASE
