+++ packages/gnars-contracts/.env.example
@@ -1,4 +1,6 @@
 DEFENDER_API_KEY=
 ETHERSCAN_API_KEY=
+BASESCAN_API_KEY=
 DEPLOYER_PRIVATE_KEY=
 JSON_RPC_URL=
+BASE_JSON_RPC_URL=
