+++ packages/webapp/.env.example
@@ -1,3 +1,4 @@
 NEXT_PUBLIC_ALCHEMY_API_KEY=
 NEXT_PUBLIC_ETHERSCAN_API_KEY=
+NEXT_PUBLIC_BASESCAN_API_KEY=
 NEXT_PUBLIC_WALLET_CONNECT_PROJECT_ID=
