+++ packages/webapp/constants/env.ts
@@ -1,3 +1,4 @@
 export const alchemyApiKey = process.env.NEXT_PUBLIC_ALCHEMY_API_KEY
 export const etherscanApiKey = process.env.NEXT_PUBLIC_ETHERSCAN_API_KEY
+export const baseScanApiKey = process.env.NEXT_PUBLIC_BASESCAN_API_KEY
 export const walletConnectProjectId = process.env.NEXT_PUBLIC_WALLET_CONNECT_PROJECT_ID!
